refactor: move app store to src root
